test(datetime-button): migrate basic
